Add function to show the edit encryption password form
showEditEncPWForm() is adapted from showForm(num) to edit and create
passwords. It binds a click event handler to the primary button, that
checks if the current password is correct and if the new password is
entered twice.

If all validations succeed, the entered password is used as the
encryption password and the users data is updated with this new
password.

Issue: https://github.com/Luzifer/cloudkeys-go/issues/10