Replace insecure password hashing
Prior this commit passwords were hashed with a static salt and using the
SHA1 hashing function. This could lead to passwords being attackable in
case someone gets access to the raw data stored inside the database.
This commit introduces password hashing using bcrypt hashing function
which addresses this issue.

Old passwords are not automatically re-hashed as they are unknown.
Replacing the old password scheme is not that easy and needs #10 to be
solved. Therefore the old hashing scheme is kept for compatibility
reason.

Signed-off-by: Knut Ahlers <knut@ahlers.me>